Add currency support to models.
The foundation of multi-currency support is going to be modifying the
code to support storing the currency in the various objects used by
Spree.

The goal is to modify the code to allow:

- adjustments
- line items
- orders
- payments
- return authorizations
- shipments
- shipping rates
- variants

to eventually store and track their own currency.

Currently these objects mirror the global currency value.  As code
support improves, this will be stored in the database for these models,
or inferred from dependent models (for example, determining the line
item currency from the order currency).