Remove the money helper from base_helper.
The money helper is based around having a single site wide currency.
This is instead being moved to a series of display_* calls on the models
which have both the monetary amount and currency to produce its own
Spree::Money object, which then allows us to display a properly
formatted currency on the output.