Remove duplicate assertions
The assertion that this item exists isn't whats being tested. There is a
spec for it above, in 'creates a carton with the shipment's inventory
units'.

If that is ever not true, that spec will fail. While this might
arguably give us better errors if that wasn't true, I don't think its
worth the duplicate expectations.

In each test we're not verifying the state of the world, just what we're
testing. We might refactor this later so that a new carton is *not* created,
or a new carton is created but an old one is deleted, but the test for the
external id is still valid.