DB migrations to backfill Carton data
The backfill migration was very slow for our database with many
shipments and inventory units so we've implemented this via raw SQL.

We also provide it as a rake task so that stores can re-run the task
if needed after deploying the new code.  See the notes in the rake
task.