Allow admin association of users to stock locations.
In the case where you have people working in physical stock locations,
you want to be able to restrict which stock locations those people have
the ability to fulfill out of. E.g. you wouldn't want the worker in your
New York store to create admin orders to fulfill out of your California
store. Adding a many-many association between users and stock
locations enables the ability to restrict which stock locations a user
has the access to do anything in.

This step just adds the association and administration, the restriction
based off of this association will come in another set of work.