Move table_filter partial within the main content in the admin layout
When adding a table filter to a page with a sidebar that table filter should be part of the twelve column wrapper div that encloses the table.

Having the table_filter outside the main content can lead to visual issues where there is not enough space for the sidebar and it’s displayed after the table rather than next the table_filter.