Fix issue with insufficient stock validation when completing an order
The current validation that checks whether there is stock for a variant when completing an order will check to see if there is stock in *any active* stock location. When adding a variant from a specific stock location, this validation is incorrect because it doesn’t check the stock in the stock location it’s meant to be shipped out of. This will also not validate correctly if the stock location is inactive because as long as there is an active stock location with stock for the variant, the validation would pass.