Allow the ability to suppress a shipment confirmation mailer.
Sometimes shipments need to be manually marked as shipped for a store
that usually automatically marks shipments as shipped. This can be due
to dropped communications with their fulfillment provider.

In these cases, marking a shipment as shipped manually is mainly for
data cleanup, and sending a shipment confirmation can create confusion.

Fix specs

Add comment

Add more specs