Enable the ability to manually mark a processing payment as failed.
The 'processing' state for payments can end up being a dead end if the
request to the payment processor fails. It is unclear when that happens
whether the charge has successfully completed, and even if it completed
successfully it does not have the proper information to mark it as a
success. This allows the resolution of payments in the processing state,
by marking them as a failure, and the person digging into the issue can
void the potential charge in the payment processor + create a new
payment here in spree.