Customer can return an item that was previously unexchanged
This fixes the StateMachine::InvalidTransition bug about calling
:receive from :unexchanged state

This represents a case when the customer asked for an exchange, and
ultimately ended up returning both the exchange and the original.
Perhaps even in the same box, but we processed the exchange item first.
It results in the creation of a new ReturnItem for the inventory unit